added the two fvec members down() for downcasting to one dimension less clip() for de-homogenization with w-clip
in drawable now exists a new function to compute a ray through a pixel specified by mouse coordinates
ray3 get_world_ray(x,y,view_ptr,modelview_projection_window_matrix)

thick_curve_viewer.cxx now supports gizmo interaction on bezier tube representation with translation and scaling of the control spheres. The new ray computation method is used to allow control sphere picking with Shift-LMB.

+++ cgv/media/transfer_function.cxx
@@ -256,7 +256,7 @@ std::vector<float> transfer_function::get_ticks(size_t request_count) const {
 	return ticks;
 	*/
 	const vec2 domain = get_domain();
-	return compute_ticks(domain[0], domain[1], request_count);
+	return compute_ticks(domain[0], domain[1], (int)request_count);
 }
 
 void transfer_function::set_interpolation(InterpolationMode interpolation) {
